extract spec_rdb
